Updating README and Creation of my Diagram
To complete this exercise (Ex 0.4) I had to learn about markdown files and languages and how they work with GitHub. I quickly found out how to write a block of markdown code in editors to get what I want, but could not figure out where to save or view markdown code. This change to the README file is the result of finding out some ways markdown languages work with GitHub and how they are saved (by the .md extentions).